test(storage): extract range_iterator inline tests to sibling file
Move RangeIterator unit tests into range_iterator_tests.rs without
changing production code or cursor-seek assertions. Preserve module
path index::range_iterator::tests and private field access.
